# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Primary Practice API
  version: 0.1.0
  license:
    name: MIT
    identifier: MIT
  summary: TOEFL Primary practice-test runner — student surface plus the admin routes v1 needs.
  description: |
    The STUDENT surface is derived from the clickable prototype at `docs/prototype/index.html`,
    screen by screen: every student endpoint exists because a screen needed it, and every field
    in a student response is consumed by one — usually rendered, sometimes control state the
    client acts on (`created`, `retryable`, `nextCursor`).

    The ADMIN routes have no prototype screens; the authoring editor is a later spec. They are
    here because v1 needs a way to get tests into the database and media onto disk.

    ## Identity
    Every route requires `Authorization: Bearer <jwt>`, verified against the JWKS at
    `auth.icovn.me`. The `sub` claim resolves to a student; a student profile IS a Google
    account, one-to-one. `POST /session` provisions from the token claims — no `GET` ever
    creates state. Admin routes additionally require the `admin` role claim.

    ## Never lose an answer
    Six rules, each earned from a real data loss in a sibling survey application:

    1. **Durable before sent.** The client writes every answer to IndexedDB on tap and clears it
       only on a per-item ack, never on request completion.
    2. **Snapshot, not delta.** `PATCH /responses` carries every answer held for ONE section.
...

    ## Ordering
...

...

...

...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /attempts/{id}/responses:
    patch:
      tags: [Attempt]
      operationId: flushResponses
      summary: Flush a section snapshot
      description: |
        The body is a FULL SNAPSHOT of every answer held for ONE section, never a delta and never
        the whole attempt. Questions in other sections are untouched, which is exactly why this is
        PATCH and not PUT. A delta loses any answer that never made it into a batch, even when
        every request sent succeeds.

        Never all-or-nothing: one rejected item cannot roll back the rest, and item failures never
        fail the envelope. Because every item shares one section, section expiry is an
        ENVELOPE-level 410 rather than a per-item reason — a per-item `section_expired` could
        never be true of some items and not others in the same request.

        Every rejected item names its capture id in `capturedAs`; nothing the server refuses is
        discarded.
      parameters:
        - { $ref: "#/components/parameters/AttemptId" }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ResponseSnapshot" }
      responses:
        "200":
          description: Per-item results. Item failures appear here, not as an HTTP error.
          content:
            {
              application/json:
                { schema: { $ref: "#/components/schemas/FlushResult" } },
            }
        "400":
          description: |
            `empty_batch`, `mixed_sections` (questions spanning more than one section), or a body
            that could not be parsed. All three are captured as a `failed_write` before the error
            is returned — see `type`.
          content:
            {
              application/problem+json:
                { schema: { $ref: "#/components/schemas/Problem" } },
            }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/NotYourAttempt" }
        "410": { $ref: "#/components/responses/SectionOrAttemptExpired" }
        "413": { $ref: "#/components/responses/PayloadCaptured" }
        "500": { $ref: "#/components/responses/UnexpectedWriteError" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Issued by auth.icovn.me; verified against its JWKS.
  parameters:
    AttemptId:
      name: id
      in: path
      required: true
      schema: { type: string }
...
  responses:
...
    Unauthorized:
      description: Missing, invalid or expired token.
      content:
        {
          application/problem+json:
            { schema: { $ref: "#/components/schemas/Problem" } },
        }
    NotYourAttempt:
      description: The attempt belongs to another student.
      content:
        {
          application/problem+json:
            { schema: { $ref: "#/components/schemas/Problem" } },
        }
...
    PayloadCaptured:
      description: |
        The body exceeded the size limit and was NOT applied. It is captured as a `failed_write`
        first, falling back to the raw bytes because an oversized body may never have been parsed.
        `capturedAs` names the row, so nothing is lost even though nothing was written.
      content:
        application/problem+json:
          schema:
            allOf:
              - { $ref: "#/components/schemas/Problem" }
              - type: object
                required: [capturedAs]
                properties:
                  capturedAs: { type: string }
...
    SectionOrAttemptExpired:
      description: |
        Two distinct conditions on one status. `section_expired` leaves the attempt running, so it
        carries no finalized attempt. `attempt_expired` means the attempt was finalized by this
        request and MUST carry it, so the client can render the time-up screen without a
        follow-up read.
      content:
        application/problem+json:
          schema:
            oneOf:
              - { $ref: "#/components/schemas/SectionExpiredProblem" }
              - { $ref: "#/components/schemas/AttemptExpiredProblem" }
    UnexpectedWriteError:
      description: |
        A database error this write did not recognise as one of its own rejection reasons — a
        deadlock, a lock timeout, a serialization failure, or anything else this API has not
        classified. Captured as a `failed_write` BEFORE this response is returned, the same as
        every other refused write (rule 4). Unlike every 4xx this API declares, `retryable` is
        true here: the identical body may well succeed on a later attempt, so the client backs
        off and retries (rule 5) rather than stopping dead.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/UnexpectedErrorProblem" }
  schemas:
    SectionExpiredProblem:
      type: object
      description: |
        The section's clock ran out; the attempt itself is still running. Closed deliberately —
        an `attempt` here would assert a finalization that did not happen.
      required: [type, title, status]
      additionalProperties: false
      properties:
        type: { type: string, const: section_expired }
        title: { type: string }
        status: { type: integer, const: 410 }
        detail: { type: string }
        retryable: { type: boolean, const: false }
    AttemptExpiredProblem:
      type: object
      description: |
        The attempt was past its deadline and has been finalized by this request. `attempt` is
        required — a client told only "expired" would have to make another call to learn what
        happened to the work.
      required: [type, title, status, attempt]
      properties:
        type: { type: string, const: attempt_expired }
        title: { type: string }
        status: { type: integer, const: 410 }
        detail: { type: string }
        retryable: { type: boolean, const: false }
        attempt: { $ref: "#/components/schemas/FinalizedAttempt" }
      additionalProperties: false
    Problem:
      type: object
      description: RFC 9457 problem details.
      required: [type, title, status]
      properties:
        type: { type: string, examples: ["attempt_expired"] }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        retryable:
          type: boolean
          const: false
          description: |
            When present, always false — every 4xx this API declares is a rule decision, not a
            transient fault, and a retried 409 is a storm rather than a recovery. A retryable
            condition (429, or any 5xx) would carry its own schema rather than relax this one.
    UnexpectedErrorProblem:
      type: object
      description: |
        The retryable 5xx `Problem`'s own schema promises — see `UnexpectedWriteError` for when
        this is returned, and `Problem`'s own `retryable` doc for why it is a separate schema
        rather than a relaxed `Problem` (`retryable` there is `const: false`, unconditionally).
      required: [type, title, status, retryable, capturedAs]
      additionalProperties: false
      properties:
        type: { type: string, const: unexpected_error }
        title: { type: string }
        status: { type: integer, const: 500 }
        detail: { type: string }
        retryable: { type: boolean, const: true }
        capturedAs: { type: string }
...
    FinalizedAttempt:
      type: object
      description: |
        An attempt finalized by lazy expiry — either the one this request touched, or a stale one
        cleared out of the way by a new start. Both uses are expiry-driven, so `status` is
        always `expired`; `submittedAt` is pinned to the deadline, never to arrival.
      required: [id, status, submittedAt, resultUrl]
      additionalProperties: false
      properties:
        id: { type: string }
        status: { type: string, const: expired }
        submittedAt: { type: string, format: date-time }
        resultUrl: { type: string }
...
    ResponseSnapshotItem:
      type: object
      required: [questionId, seq, selectedChoiceIds]
      additionalProperties: false
      properties:
        questionId: { type: string }
        seq: { type: integer }
        selectedChoiceIds:
          type: array
          items: { type: string }
        answeredAt: { type: string, format: date-time }
        timeSpentMs: { type: integer }
    ResponseSnapshot:
      type: object
      required: [clientInstanceId, responses]
      additionalProperties: false
      properties:
        clientInstanceId: { type: string }
        responses:
          type: array
          minItems: 1
          items: { $ref: "#/components/schemas/ResponseSnapshotItem" }
          description: |
            Every answer held for ONE section — a snapshot, never a delta. A body whose questions
            span more than one section is refused as `mixed_sections` and captured, because the
            envelope-level section rules cannot be evaluated against two sections at once.
    ItemAccepted:
      type: object
      description: The write landed, or was discarded as stale. Both are successes.
      required: [questionId, status]
      additionalProperties: false
      properties:
        questionId: { type: string }
        status: { type: string, enum: [applied, ignored_stale] }
    ItemRejected:
      type: object
      description: |
        The write was refused. `capturedAs` is REQUIRED: a rejection without a capture id is an
        answer the server refused and did not keep, which is the failure mode this whole design
        exists to prevent.
      required: [questionId, status, reason, retryable, capturedAs]
      additionalProperties: false
      properties:
        questionId: { type: string }
        status: { type: string, const: rejected }
        reason:
          type: string
          enum:
            [
              answer_change_not_allowed,
              navigation_locked,
              unknown_question,
              invalid,
            ]
          description: Per-item conditions only. `section_expired` and `mixed_sections` are envelope-level.
        retryable:
          type: boolean
          const: false
          description: |
            Always false, and enforced as such. Every per-item rejection is a rule decision, not a
            transient fault — a retried 409-class failure is a storm, not a recovery.
        capturedAs:
          type: string
          description: The `failed_write` id holding this item verbatim.
    ItemResult:
      oneOf:
        - { $ref: "#/components/schemas/ItemAccepted" }
        - { $ref: "#/components/schemas/ItemRejected" }
      discriminator:
        propertyName: status
        mapping:
          applied: "#/components/schemas/ItemAccepted"
          ignored_stale: "#/components/schemas/ItemAccepted"
          rejected: "#/components/schemas/ItemRejected"
    AttemptClock:
      type: object
      required: [serverTime]
      properties:
        expiresAt: { type: [string, "null"], format: date-time }
        sectionExpiresAt: { type: [string, "null"], format: date-time }
        serverTime: { type: string, format: date-time }
...
    FlushResult:
      type: object
      required: [results, attempt]
      properties:
        results:
          type: array
          items: { $ref: "#/components/schemas/ItemResult" }
          description: One entry per snapshot item — the client clears its queue per ack, not per request.
        attempt: { $ref: "#/components/schemas/AttemptClock" }
...